Handle the number of updated rows as a Long value (#581)
* Handle the number of updated rows as a Long value.

Because R2DBC 1.0 spec does it.

* Polish

+++ integration-test-r2dbc/src/test/kotlin/integration/r2dbc/R2dbcInsertBatchTest.kt
@@ -101,9 +101,9 @@ class R2dbcInsertBatchTest(private val db: R2dbcDatabase) {
         val query = QueryDsl.insert(d).onDuplicateKeyUpdate().batch(department1, department2)
         val counts = db.runQuery { query }
         when (db.config.dialect.driver) {
-            "mariadb" -> assertEquals(listOf(Statement.SUCCESS_NO_INFO, Statement.SUCCESS_NO_INFO), counts)
-            "mysql" -> assertEquals(listOf(1, 2), counts)
-            else -> assertEquals(listOf(1, 1), counts)
+            "mariadb" -> assertEquals(listOf(Statement.SUCCESS_NO_INFO.toLong(), Statement.SUCCESS_NO_INFO.toLong()), counts)
+            "mysql" -> assertEquals(listOf(1L, 2L), counts)
+            else -> assertEquals(listOf(1L, 1L), counts)
         }
         val list = db.runQuery {
             QueryDsl.from(d).where { d.departmentId inList listOf(1, 5) }.orderBy(d.departmentId)
@@ -124,9 +124,9 @@ class R2dbcInsertBatchTest(private val db: R2dbcDatabase) {
         val query = QueryDsl.insert(d).onDuplicateKeyUpdate(d.departmentNo).batch(department1, department2)
         val counts = db.runQuery { query }
         when (db.config.dialect.driver) {
-            "mariadb" -> assertEquals(listOf(Statement.SUCCESS_NO_INFO, Statement.SUCCESS_NO_INFO), counts)
-            "mysql" -> assertEquals(listOf(1, 2), counts)
-            else -> assertEquals(listOf(1, 1), counts)
+            "mariadb" -> assertEquals(listOf(Statement.SUCCESS_NO_INFO.toLong(), Statement.SUCCESS_NO_INFO.toLong()), counts)
+            "mysql" -> assertEquals(listOf(1L, 2L), counts)
+            else -> assertEquals(listOf(1L, 1L), counts)
         }
         val list = db.runQuery {
             QueryDsl.from(d).where { d.departmentNo inList listOf(10, 50) }.orderBy(d.departmentNo)
@@ -150,8 +150,8 @@ class R2dbcInsertBatchTest(private val db: R2dbcDatabase) {
             }.batch(listOf(department1, department2))
         val counts = db.runQuery { query }
         when (db.config.dialect.driver) {
-            "mysql", "mariadb" -> assertEquals(listOf(1, 2), counts)
-            else -> assertEquals(listOf(1, 1), counts)
+            "mysql", "mariadb" -> assertEquals(listOf(1L, 2L), counts)
+            else -> assertEquals(listOf(1L, 1L), counts)
         }
         val list = db.runQuery {
             QueryDsl.from(d).where { d.departmentId inList listOf(1, 5) }.orderBy(d.departmentId)
@@ -175,8 +175,8 @@ class R2dbcInsertBatchTest(private val db: R2dbcDatabase) {
             }.batch(department1, department2)
         val counts = db.runQuery { query }
         when (db.config.dialect.driver) {
-            "mysql", "mariadb" -> assertEquals(listOf(1, 2), counts)
-            else -> assertEquals(listOf(1, 1), counts)
+            "mysql", "mariadb" -> assertEquals(listOf(1L, 2L), counts)
+            else -> assertEquals(listOf(1L, 1L), counts)
         }
         val list = db.runQuery {
             QueryDsl.from(d).where { d.departmentNo inList listOf(10, 50) }.orderBy(d.departmentNo)
@@ -197,8 +197,8 @@ class R2dbcInsertBatchTest(private val db: R2dbcDatabase) {
         val query = QueryDsl.insert(d).onDuplicateKeyIgnore().batch(listOf(department1, department2))
         val counts = db.runQuery { query }
         when (db.config.dialect.driver) {
-            "mariadb" -> assertEquals(listOf(Statement.SUCCESS_NO_INFO, Statement.SUCCESS_NO_INFO), counts)
-            else -> assertEquals(listOf(1, 0), counts)
+            "mariadb" -> assertEquals(listOf(Statement.SUCCESS_NO_INFO.toLong(), Statement.SUCCESS_NO_INFO.toLong()), counts)
+            else -> assertEquals(listOf(1L, 0L), counts)
         }
         val list = db.runQuery {
             QueryDsl.from(d).where { d.departmentId inList listOf(1, 5) }.orderBy(d.departmentId)
@@ -219,8 +219,8 @@ class R2dbcInsertBatchTest(private val db: R2dbcDatabase) {
         val query = QueryDsl.insert(d).onDuplicateKeyIgnore(d.departmentNo).batch(listOf(department1, department2))
         val counts = db.runQuery { query }
         when (db.config.dialect.driver) {
-            "mariadb" -> assertEquals(listOf(Statement.SUCCESS_NO_INFO, Statement.SUCCESS_NO_INFO), counts)
-            else -> assertEquals(listOf(1, 0), counts)
+            "mariadb" -> assertEquals(listOf(Statement.SUCCESS_NO_INFO.toLong(), Statement.SUCCESS_NO_INFO.toLong()), counts)
+            else -> assertEquals(listOf(1L, 0L), counts)
         }
         val list = db.runQuery {
             QueryDsl.from(d).where { d.departmentNo inList listOf(10, 50) }.orderBy(d.departmentNo)
