+++ komapper-core/src/main/kotlin/org/komapper/core/dsl/query/InsertOnDuplicateKeyIgnoreQueryBuilder.kt
@@ -19,7 +19,7 @@ interface InsertOnDuplicateKeyIgnoreQueryBuilder<ENTITY : Any, ID : Any, META : 
      * @param entity the entity to be inserted
      * @return the query
      */
-    fun single(entity: ENTITY): EntityUpsertQuery<Int>
+    fun single(entity: ENTITY): EntityUpsertQuery<Long>
 
     /**
      * Builds a query to bulk insert a list of entities.
@@ -27,7 +27,7 @@ interface InsertOnDuplicateKeyIgnoreQueryBuilder<ENTITY : Any, ID : Any, META : 
      * @param entities the entities to be inserted
      * @return the query
      */
-    fun multiple(entities: List<ENTITY>): EntityUpsertQuery<Int>
+    fun multiple(entities: List<ENTITY>): EntityUpsertQuery<Long>
 
     /**
      * Builds a query to bulk insert an array of entities.
@@ -35,7 +35,7 @@ interface InsertOnDuplicateKeyIgnoreQueryBuilder<ENTITY : Any, ID : Any, META : 
      * @param entities the entities to be inserted
      * @return the query
      */
-    fun multiple(vararg entities: ENTITY): EntityUpsertQuery<Int>
+    fun multiple(vararg entities: ENTITY): EntityUpsertQuery<Long>
 
     /**
      * Builds a query to insert a list of entities in a batch.
@@ -43,7 +43,7 @@ interface InsertOnDuplicateKeyIgnoreQueryBuilder<ENTITY : Any, ID : Any, META : 
      * @param entities the entities to be inserted
      * @return the query
      */
-    fun batch(entities: List<ENTITY>, batchSize: Int? = null): EntityUpsertQuery<List<Int>>
+    fun batch(entities: List<ENTITY>, batchSize: Int? = null): EntityUpsertQuery<List<Long>>
 
     /**
      * Builds a query to insert an array of entities in a batch.
@@ -51,7 +51,7 @@ interface InsertOnDuplicateKeyIgnoreQueryBuilder<ENTITY : Any, ID : Any, META : 
      * @param entities the entities to be inserted
      * @return the query
      */
-    fun batch(vararg entities: ENTITY, batchSize: Int? = null): EntityUpsertQuery<List<Int>>
+    fun batch(vararg entities: ENTITY, batchSize: Int? = null): EntityUpsertQuery<List<Long>>
 
     /**
      * Builds a query to insert or update a single entity and get the result as a new entity.
@@ -68,23 +68,23 @@ internal data class InsertOnDuplicateKeyIgnoreQueryBuilderImpl<ENTITY : Any, ID 
 
     private val builder: EntityUpsertQueryBuilder<ENTITY, ID, META> = EntityUpsertQueryBuilderImpl(context)
 
-    override fun single(entity: ENTITY): EntityUpsertQuery<Int> {
+    override fun single(entity: ENTITY): EntityUpsertQuery<Long> {
         return builder.single(entity)
     }
 
-    override fun multiple(entities: List<ENTITY>): EntityUpsertQuery<Int> {
+    override fun multiple(entities: List<ENTITY>): EntityUpsertQuery<Long> {
         return builder.multiple(entities)
     }
 
-    override fun multiple(vararg entities: ENTITY): EntityUpsertQuery<Int> {
+    override fun multiple(vararg entities: ENTITY): EntityUpsertQuery<Long> {
         return builder.multiple(entities.toList())
     }
 
-    override fun batch(entities: List<ENTITY>, batchSize: Int?): EntityUpsertQuery<List<Int>> {
+    override fun batch(entities: List<ENTITY>, batchSize: Int?): EntityUpsertQuery<List<Long>> {
         return builder.batch(entities, batchSize)
     }
 
-    override fun batch(vararg entities: ENTITY, batchSize: Int?): EntityUpsertQuery<List<Int>> {
+    override fun batch(vararg entities: ENTITY, batchSize: Int?): EntityUpsertQuery<List<Long>> {
         return builder.batch(entities.toList(), batchSize)
     }
